+++ src/explorer/macroExplorerTreeItems.ts
@@ -133,9 +133,9 @@ export function getRunItem(runInfo: MacroRunInfo) {
     return isSameDay
       ? `Started at ${time}`
       : `Started on ${date.toLocaleDateString([], {
-        year: '2-digit',
-        month: 'numeric',
-        day: 'numeric',
-      })}, ${time}`;
+          year: '2-digit',
+          month: 'numeric',
+          day: 'numeric',
+        })}, ${time}`;
   }
 }